+++ drizzle/resample.py
@@ -167,8 +167,8 @@ class Drizzle:
 
     def __init__(self, kernel="square", fillval=None, fillval2=None,
                  out_shape=None, out_img=None, out_wht=None, out_ctx=None,
-                 out_img2=None, exptime=0.0, begin_ctx_id=0, max_ctx_id=None,
-                 disable_ctx=False):
+                 out_img2=None, out_dq=None, exptime=0.0, begin_ctx_id=0,
+                 max_ctx_id=None, disable_ctx=False):
         """
         kernel: str, optional
             The name of the kernel used to combine the input. The choice of
@@ -232,6 +232,14 @@ class Drizzle:
             Subsequent calls hold intermediate results. This parameter is
             ignored when ``disable_ctx`` is `True`.
 
+        out_dq : 2D array of int32, None, optional
+            A 2D `~numpy.ndarray` containing DQ bitfields of output (resampled)
+            pixels. It is computed by combining (using bitwise-OR) DQ bitfields
+            of input pixels that contributed to the output pixel.
+            If provided, it must be a 2D array of the same shape as
+            ``out_img``. If `None`, output DQ array will be created during
+            the first call to `add_image` and will be initialized to zero.
+
         exptime : float, optional
             Exposure time of previously resampled images when provided via
             parameters ``out_img``, ``out_wht``, ``out_ctx``.
@@ -262,6 +270,7 @@ class Drizzle:
         """
         self._ncoadds = 0
         self._out_img2 = None
+        self._out_dq = None
         self._disable_ctx = disable_ctx
 
         if disable_ctx:
@@ -328,6 +337,10 @@ class Drizzle:
                 raise ValueError("'out_ctx' must be either a 2D or 3D array.")
             shapes.add(out_ctx.shape[1:])
 
+        if out_dq is not None:
+            out_dq = np.asarray(out_dq, dtype=np.int32)
+            shapes.add(out_dq.shape)
+
         if out_shape is not None:
             shapes.add(tuple(out_shape))
 
@@ -344,7 +357,8 @@ class Drizzle:
         elif len(shapes) > 1:
             raise ValueError(
                 "Inconsistent data shapes specified: 'out_shape' and/or "
-                "out_img, out_img2, out_wht, out_ctx have different shapes."
+                "out_img, out_img2, out_wht, out_ctx, out_dq have different "
+                "shapes."
             )
         else:
             self._out_shape = None
@@ -370,6 +384,14 @@ class Drizzle:
                 )
         self._output_shapes = shapes
         self._alloc_output_arrays2_init(out_img2=out_img2)
+
+        if out_dq is not None:
+            out_dq = np.asarray(out_dq, dtype=np.int32)
+            if out_dq.shape != self._out_shape:
+                raise ValueError(
+                    "'out_dq' must have the same shape as 'out_img'."
+                )
+            self._out_dq = out_dq
 
     @property
     def fillval(self):
@@ -416,6 +438,14 @@ class Drizzle:
 
         """
         return self._out_img2
+
+    @property
+    def out_dq(self):
+        """Output DQ image computed by OR-combining DQ bitfields of input
+        images' pixels that have contributed to a given ouput pixel.
+
+        """
+        return self._out_dq
 
     @property
     def total_exptime(self):
@@ -563,7 +593,7 @@ class Drizzle:
 
         return plane_info
 
-    def add_image(self, data, exptime, pixmap, data2=None, scale=1.0,
+    def add_image(self, data, exptime, pixmap, data2=None, dq=None, scale=1.0,
                   weight_map=None, wht_scale=1.0, pixfrac=1.0, in_units='cps',
                   xmin=None, xmax=None, ymin=None, ymax=None):
         """
@@ -606,6 +636,14 @@ class Drizzle:
                 ``data``. Therefore, when ``in_units`` are "counts",
                 ``data2`` arrays will be rescaled by ``exptime**2`` to convert
                 to rate units before resampling.
+
+        dq : 2D array, None, optional
+            A 2D numpy array containing DQ bitfields of input pixels. It must
+            have the same shape as ``data``. If provided, output DQ array
+            (accessible via ``out_dq`` property) will be computed by combining
+            (using bitwise-OR) DQ bitfields of input pixels that contributed to
+            the output pixel. If `None`, DQ array of the output image will
+            not be computed.
 
         scale : float, optional
             The pixel scale of the input image. Conceptually, this is the
@@ -786,6 +824,15 @@ class Drizzle:
                 raise AssertionError("Context image is expected to be 3D")
             ctx_plane = self._out_ctx[plane_no]
 
+        if dq is not None:
+            dq = np.asarray(dq, dtype=np.int32)
+            if dq.shape != data.shape:
+                raise ValueError(
+                    "'dq' shape is not consistent with 'data' shape."
+                )
+            if self._out_dq is None:
+                self._out_dq = np.zeros(self._out_shape, dtype=np.int32)
+
         # TODO: probably tdriz should be modified to not return version.
         #       we should not have git, Python, C, ... versions
 
@@ -795,13 +842,15 @@ class Drizzle:
 
         _vers, nmiss, nskip = cdrizzle.tdriz(
             input=data,
-            input2=data2,
             weights=weight_map,
             pixmap=pixmap,
             output=self._out_img,
-            output2=self._out_img2,
             counts=self._out_wht,
             context=ctx_plane,
+            input2=data2,
+            output2=self._out_img2,
+            dq=dq,
+            outdq=self._out_dq,
             uniqid=id_in_plane + 1,
             xmin=xmin,
             xmax=xmax,