+++ src/cdrizzlebox.c
@@ -86,11 +86,12 @@ update_data(struct driz_param_t *p, const integer_t ii, const integer_t jj,
  */
 inline_macro static int
 update_data_var(struct driz_param_t *p, const integer_t ii, const integer_t jj,
-                const float d, const float dow, float *d2) {
+                const float d, const float dow, float *d2, int dq) {
     double vc_plus_dow, vc_plus_dow2;
     double v, vc2, dow2;
     float vc;
     int i;
+    int output_dq_value;
     PyArrayObject **arr2;
 
     if (oob_output_pixel(p, ii, jj)) {
@@ -116,6 +117,9 @@ update_data_var(struct driz_param_t *p, const integer_t ii, const integer_t jj,
                 set_pixel(arr2[i], ii, jj, d2[i]);
             }
         }
+        if (p->dq) {
+            set_int_pixel(p->output_dq, ii, jj, dq);
+        }
     } else {
         v = (get_pixel(p->output_data, ii, jj) * vc + dow * d) / vc_plus_dow;
         set_pixel(p->output_data, ii, jj, v);
@@ -128,6 +132,10 @@ update_data_var(struct driz_param_t *p, const integer_t ii, const integer_t jj,
                     vc_plus_dow2;
                 set_pixel(arr2[i], ii, jj, v);
             }
+        }
+        if (p->dq) {
+            output_dq_value = get_int_pixel(p->output_dq, ii, jj);
+            set_int_pixel(p->output_dq, ii, jj, output_dq_value | dq);
         }
     }
     set_pixel(p->output_counts, ii, jj, vc_plus_dow);
@@ -474,7 +482,7 @@ do_kernel_point_var(struct driz_param_t *p) {
     integer_t osize[2];
     float scale2, scale4, d, dow, *d2 = NULL;
     integer_t bv;
-    int xmin, xmax, ymin, ymax, n;
+    int xmin, xmax, ymin, ymax, n, dqval = 0;
     int ndata2;
 
     ndata2 = p->ndata2;
@@ -573,13 +581,17 @@ do_kernel_point_var(struct driz_param_t *p) {
                         dow = 1.0f;
                     }
 
+                    if (p->dq) {
+                        dqval = (int)get_int_pixel(p->dq, i, j);
+                    }
+
                     /* If we are creating or modifying the context image,
                        we do so here. */
                     if (p->output_context && dow > 0.0f) {
                         set_bit(p->output_context, ii, jj, bv);
                     }
 
-                    if (update_data_var(p, ii, jj, d, dow, d2)) {
+                    if (update_data_var(p, ii, jj, d, dow, d2, dqval)) {
                         free(d2);
                         return 1;
                     }
@@ -608,7 +620,7 @@ do_kernel_gaussian_var(struct driz_param_t *p) {
     double gaussian_efac, gaussian_es;
     double pfo, ac, scale2, scale4, w, ddx, ddy, r2, dover;
     const double nsig = 2.5;
-    int xmin, xmax, ymin, ymax, n;
+    int xmin, xmax, ymin, ymax, n, dqval = 0;
     int ndata2;
 
     ndata2 = p->ndata2;
@@ -739,7 +751,11 @@ do_kernel_gaussian_var(struct driz_param_t *p) {
                             set_bit(p->output_context, ii, jj, bv);
                         }
 
-                        if (update_data_var(p, ii, jj, d, dow, d2)) {
+                        if (p->dq) {
+                            dqval = (int)get_int_pixel(p->dq, i, j);
+                        }
+
+                        if (update_data_var(p, ii, jj, d, dow, d2, dqval)) {
                             free(d2);
                             return 1;
                         }
@@ -777,7 +793,7 @@ do_kernel_lanczos_var(struct driz_param_t *p) {
     size_t ix, iy;
     double sdp;
     double *lut = NULL;
-    int xmin, xmax, ymin, ymax, n;
+    int xmin, xmax, ymin, ymax, n, dqval = 0;
     int ndata2;
 
     if (fabs(p->pixel_fraction - 1.0) > 1.0e-5) {
@@ -919,7 +935,11 @@ do_kernel_lanczos_var(struct driz_param_t *p) {
                             set_bit(p->output_context, ii, jj, bv);
                         }
 
-                        if (update_data_var(p, ii, jj, d, dow, d2)) {
+                        if (p->dq) {
+                            dqval = (int)get_int_pixel(p->dq, i, j);
+                        }
+
+                        if (update_data_var(p, ii, jj, d, dow, d2, dqval)) {
                             free(d2);
                             free(lut);
                             return 1;
@@ -957,7 +977,7 @@ do_kernel_turbo_var(struct driz_param_t *p) {
     float d, dow, *d2 = NULL;
     double pfo, scale2, scale4, ac;
     double xxi, xxa, yyi, yya, w, dover;
-    int xmin, xmax, ymin, ymax, n;
+    int xmin, xmax, ymin, ymax, n, dqval = 0;
     int ndata2;
 
     ndata2 = p->ndata2;
@@ -1084,7 +1104,11 @@ do_kernel_turbo_var(struct driz_param_t *p) {
                                 set_bit(p->output_context, ii, jj, bv);
                             }
 
-                            if (update_data_var(p, ii, jj, d, dow, d2)) {
+                            if (p->dq) {
+                                dqval = get_int_pixel(p->dq, i, j);
+                            }
+
+                            if (update_data_var(p, ii, jj, d, dow, d2, dqval)) {
                                 return 1;
                             }
                         }
@@ -1120,7 +1144,7 @@ do_kernel_square_var(struct driz_param_t *p) {
     double dh, jaco, dover, w;
 
     double xin[4], yin[4], xout[4], yout[4];
-    int ndata2;
+    int ndata2, dqval = 0;
 
     ndata2 = p->ndata2;
 
@@ -1284,7 +1308,11 @@ do_kernel_square_var(struct driz_param_t *p) {
                             set_bit(p->output_context, ii, jj, bv);
                         }
 
-                        if (update_data_var(p, ii, jj, d, dow, d2)) {
+                        if (p->dq) {
+                            dqval = (int)get_int_pixel(p->dq, i, j);
+                        }
+
+                        if (update_data_var(p, ii, jj, d, dow, d2, dqval)) {
                             free(d2);
                             return 1;
                         }
@@ -1982,7 +2010,7 @@ dobox(struct driz_param_t *p) {
 
     /* Set up a function pointer to handle the appropriate kernel */
     if (p->kernel < kernel_LAST) {
-        if (p->ndata2 > 0) {
+        if (p->ndata2 > 0 || p->dq != NULL) {
             kernel_handler = kernel_var_handler_map[p->kernel];
         } else {
             kernel_handler = kernel_handler_map[p->kernel];